Clean up

+++ src/PixelDissimilarity.cpp
@@ -2,41 +2,41 @@
 
 template <class Pixel> float PixelDissimilarity<Pixel>::L1(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((float)_image[index1] - (float)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
-    float dist = 0.0;
+    double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
         ImgIdx offset = ch * _imageSize;
-        float diff = std::abs((float)_image[index1 + offset] - (float)_image[index2 + offset]);
+        double diff = std::abs((double)_image[index1 + offset] - (double)_image[index2 + offset]);
         dist += diff;
     }
-    return dist / (float)_channels;
+    return (float)(dist / (double)_channels);
 }
 
 template <class Pixel> float PixelDissimilarity<Pixel>::L2(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((float)_image[index1] - (float)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
-    float dist = 0.0;
+    double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
         ImgIdx offset = ch * _imageSize;
-        float diff = _image[index1 + offset] - _image[index2 + offset];
+        double diff = _image[index1 + offset] - _image[index2 + offset];
         dist += diff * diff;
     }
-    return std::sqrt(dist / (float)_channels);
+    return (float)std::sqrt(dist / (double)_channels);
 }
 
 template <class Pixel> float PixelDissimilarity<Pixel>::LInfinity(ImgIdx index1, ImgIdx index2) const {
     if (_channels == 1)
-        return std::abs((float)_image[index1] - (float)_image[index2]);
+        return (float)std::abs((double)_image[index1] - (double)_image[index2]);
 
-    float dist = 0.0;
+    double dist = 0.0;
     for (ImgIdx ch = 0; ch < _channels; ch++) {
         ImgIdx offset = ch * _imageSize;
-        float diff = std::abs((float)_image[index1 + offset] - (float)_image[index2 + offset]);
+        double diff = std::abs((double)_image[index1 + offset] - (double)_image[index2 + offset]);
         dist = std::max(dist, diff);
     }
-    return dist;
+    return (float)dist;
 }
 
 template class PixelDissimilarity<uint8_t>;
